+++ src/components/instant-apps-control-panel/instant-apps-control-panel.tsx
@@ -64,6 +64,7 @@ export class InstantAppsControlPanel {
       if (component.isExpand) {
         content = this._getExpand(component);
         if (component.expandIconClass) content.expandIconClass = component.expandIconClass;
+        if (component.collapseTooltip) content.collapseTooltip = component.collapseTooltip;
         if (component.expandTooltip) content.expandTooltip = component.expandTooltip;
       }